refs #15: Fix travis build, fix spring starter Travis builds were running mvn test and not mvn verify. This resulted in breakages in Spring Starter not being discovered.
