+++ .travis.yml
@@ -1,1 +1,2 @@
 language: java
+script: mvn clean verify
