+++ examples/java-dynamic-sqs-listener-spring-aws-example/pom.xml
@@ -9,10 +9,11 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>java-dynamic-sqs-listener-spring-starter-examples</artifactId>
+    <artifactId>java-dynamic-sqs-listener-spring-aws-example</artifactId>
 
-    <name>Java Dynamic SQS Listener - Spring Starter - Examples</name>
-    <description>Contains examples for using the Spring Starter implementation of the framework.</description>
+
+    <name>Java Dynamic SQS Listener - Spring Starter - AWS Example</name>
+    <description>Contains examples for </description>
 
     <properties>
         <findbugs.config.location>../../configuration/findbugs/bugsExcludeFilter.xml</findbugs.config.location>
