+++ examples/java-dynamic-sqs-listener-spring-aws-example/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+sqs:
+  queues:
+    queueUrl: ${SQS_QUEUE_URL}
