+++ examples/pom.xml
@@ -23,6 +23,7 @@
     <modules>
         <module>java-dynamic-sqs-listener-core-examples</module>
         <module>java-dynamic-sqs-listener-spring-starter-examples</module>
+        <module>java-dynamic-sqs-listener-spring-aws-example</module>
     </modules>
 
     <dependencies>
