--- java-dynamic-sqs-listener-spring/java-dynamic-sqs-listener-spring-starter/README.md
@@ -1,6 +1,0 @@
-# Java Dynamic SQS Listener - Spring Starter
-Starter that will auto configure this library to the spring environment when the `@EnableAutoConfiguration` annotation has been applied to your
-application.  Note that if you are running in a Spring Boot application and have the `@SpringBootApplication` annotation you do
-not need to put the `@EnableAutoConfiguration` annotation as it is already applied.
-
-For the actual Spring configuration see the [java-dynamic-sqs-listener-spring-core](../java-dynamic-sqs-listener-spring-core) module.
