--- java-dynamic-sqs-listener-spring/java-dynamic-sqs-listener-spring-starter/src/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.jashmore.sqs.spring.config.QueueListenerAutoConfiguration
