+++ java-dynamic-sqs-listener-spring/java-dynamic-sqs-listener-spring-starter/src/test/java/com/jashmore/sqs/spring/queue/DefaultQueueResolverServiceTest.java
