+++ java-dynamic-sqs-listener-spring/java-dynamic-sqs-listener-spring-starter/src/test/java/it/com/jashmore/sqs/container/basic/QueueListenerWrapperIntegrationTest.java
